# docs/openapi.yaml
openapi: 3.0.3
info:
  title: SafePay API
  version: "1.0.0"
  description: |
    **Trusted payments, everywhere.**

    SafePay is an escrow, settlement and trust-scoring layer for Nigeria. Any
    application — a marketplace, a freelance platform, a storefront, a landlord's
    portal — can plug in so that a transaction between two people who do not fully
    trust each other becomes safe by default.

    ## Authentication

    Two credentials reach the same endpoints:

    | Credential | Header | Used by |
    |---|---|---|
    | Dashboard session (JWT) | `Authorization: Bearer eyJ…` | The SafePay web app |
    | API key | `Authorization: Bearer sk_test_…` | Partner applications |

    Keys come in pairs. `sk_test_` keys operate on simulated settlement — build the
    entire flow, including disputes and webhooks, without moving a naira.
    SafePay stores only a SHA-256 hash of every key, so a lost key is rotated,
    never recovered.

    ## Money

    Every amount crosses the wire as an **integer in kobo** (`amountKobo`). Endpoints
    that create an escrow also accept `amount` in naira as a convenience, and convert
    it. Floating point never touches a balance.

    ## Errors

    Failures return a consistent envelope:

    ```json
    { "error": { "code": "conflict", "message": "An escrow that is \"released\" cannot move to \"released\"." } }
    ```

    ## Rate limits

    Responses carry `RateLimit-Limit`, `RateLimit-Remaining` and `RateLimit-Reset`.
...
  contact:
...
servers:
...
security:
  - bearerAuth: []
paths:
...
  /v1/auth/signup:
    post:
      tags: [Auth]
      summary: Create an account
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name, email, password]
              properties:
                name: { type: string, example: Ada Okonkwo }
                email: { type: string, format: email, example: ada@example.com }
                password: { type: string, minLength: 8, example: correct-horse-battery }
                phone: { type: string, example: "+2348012345678" }
      responses:
        "201":
          description: Account created; a session token is returned.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/AuthResponse" }
        "409": { $ref: "#/components/responses/Conflict" }
        "400": { $ref: "#/components/responses/BadRequest" }
  /v1/auth/login:
    post:
      tags: [Auth]
      summary: Sign in
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password]
              properties:
                email: { type: string, format: email }
                password: { type: string }
      responses:
        "200":
          description: Signed in.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/AuthResponse" }
        "401": { $ref: "#/components/responses/Unauthorized" }
  /v1/auth/me:
    get:
      tags: [Auth]
      summary: The signed-in account and its SafeScore
      responses:
        "200":
          description: Account with full score breakdown.
          content:
            application/json:
              schema:
                type: object
                properties:
                  user: { $ref: "#/components/schemas/User" }
                  score: { $ref: "#/components/schemas/Score" }
        "401": { $ref: "#/components/responses/Unauthorized" }
    patch:
      tags: [Auth]
      summary: Update the signed-in account
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                name: { type: string }
                phone: { type: string }
                verificationTier:
                  type: string
                  enum: [none, phone, bvn_nin, address]
      responses:
        "200":
          description: Updated, with the recalculated score.
          content:
            application/json:
              schema:
                type: object
                properties:
                  user: { $ref: "#/components/schemas/User" }
                  score: { $ref: "#/components/schemas/Score" }
...
  /v1/developer/apps:
    post:
      tags: [Developer]
      summary: Create an app and issue API keys
      description: The **only** time raw keys are returned. SafePay stores hashes.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name]
              properties:
                name: { type: string }
                description: { type: string }
                webhookUrl: { type: string, format: uri }
      responses:
        "201":
          description: App created with its key pair.
          content:
            application/json:
              schema:
                type: object
                properties:
                  app: { $ref: "#/components/schemas/App" }
                  keys:
                    type: object
                    properties:
                      test: { type: string, example: sk_test_AhpZn4cEtTA_IE16mui_MwQdrqvzPY-8 }
                      live: { type: string }
                  notice: { type: string }
    get:
      tags: [Developer]
      summary: List your apps
      responses:
        "200":
          description: Apps owned by the caller.
  /v1/developer/apps/{id}/rotate:
    post:
      tags: [Developer]
      summary: Rotate an API key
      description: The previous key stops working immediately.
      parameters: [{ name: id, in: path, required: true, schema: { type: string } }]
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                mode: { type: string, enum: [test, live], default: test }
      responses:
        "200":
          description: A new key, shown once.
  /v1/developer/apps/{id}/webhooks:
    get:
      tags: [Developer]
      summary: Webhook delivery log
      parameters: [{ name: id, in: path, required: true, schema: { type: string } }]
      responses:
        "200":
          description: Recent deliveries with attempt counts and errors.
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: A dashboard session JWT, or an API key beginning `sk_test_` / `sk_live_`.
...
  responses:
...
    BadRequest:
      description: The request was malformed or failed validation.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    Unauthorized:
      description: Missing or invalid credentials.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
...
    Conflict:
      description: The resource is not in a state that allows this transition.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
  schemas:
    Error:
      type: object
      properties:
        error:
          type: object
          properties:
            code: { type: string, example: conflict }
            message: { type: string }
            details: { type: object }
    User:
      type: object
      properties:
        id: { type: string, example: usr_JbRmBoCmk1JrSKft }
        name: { type: string }
        email: { type: string, format: email }
        phone: { type: string, nullable: true }
        role: { type: string, enum: [user, admin] }
        verificationTier: { type: string, enum: [none, phone, bvn_nin, address] }
        safeScore: { type: integer, example: 68 }
        scoreTier: { type: string, enum: [new, building, trusted, verified_pro] }
        createdAt: { type: string, format: date-time }
    AuthResponse:
      type: object
      properties:
        token: { type: string, description: Bearer this on subsequent requests. }
        user: { $ref: "#/components/schemas/User" }
...
    Score:
      type: object
      description: |
        The full, explainable score. Components are additive and bounded, so the
        number can always be interrogated — a trust score nobody can audit is a
        trust score nobody trusts.
      properties:
        userId: { type: string }
        score: { type: integer, minimum: 0, maximum: 100, example: 68 }
        tier: { type: string, enum: [new, building, trusted, verified_pro] }
        tierLabel: { type: string, example: Trusted }
        breakdown:
          type: object
          properties:
            volume: { type: number, description: "Completed deals (max 25)." }
            value: { type: number, description: "Value settled (max 10)." }
            reliability: { type: number, description: "Dispute-free record, scaled by evidence (max 25)." }
            speed: { type: number, description: "Median time to confirm (max 10)." }
            verification: { type: number, description: "Identity assurance tier (max 15)." }
            age: { type: number, description: "Account tenure (max 5)." }
            diversity: { type: number, description: "Distinct counterparties — anti-gaming (max 10)." }
        weights: { type: object }
        stats:
          type: object
          properties:
            escrowsCompleted: { type: integer }
            disputeRatePct: { type: number }
            uniqueCounterparties: { type: integer }
            medianReleaseHours: { type: number, nullable: true }
            totalValueSettledNaira: { type: integer }
            accountAgeDays: { type: integer }
            concentrationFlag:
              type: boolean
              description: True when volume is concentrated in too few counterparties; the score is capped.
...
    App:
      type: object
      properties:
        id: { type: string }
        name: { type: string }
        description: { type: string }
        webhookUrl: { type: string, nullable: true }
        webhookSecret: { type: string, example: whsec_… }
        subscribedEvents:
          type: array
          items: { type: string }
        testKeyPreview: { type: string, description: Masked. The raw key is shown only at creation or rotation. }
        liveKeyPreview: { type: string }
        liveEnabled: { type: boolean }
        createdAt: { type: string, format: date-time }
...